backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
motor==3.3.1
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
